+++ auditor.core/src/main/java/com/github/tkaczenko/auditor/core/EnableAuditor.java
@@ -1,6 +1,6 @@
 package com.github.tkaczenko.auditor.core;
 
-import com.github.tkaczenko.auditor.core.config.AuditorConfiguration;
+import com.github.tkaczenko.auditor.core.config.AuditorConfig;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -10,9 +10,9 @@ import org.springframework.context.annotation.Import;
 /**
  * This annotation is used to enable the Auditor functionality in a Spring application. When applied
  * to the main application class or a configuration class, it will import the necessary {@link
- * AuditorConfiguration} to set up the Auditor functionality.
+ * AuditorConfig} to set up the Auditor functionality.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-@Import(AuditorConfiguration.class)
+@Import(AuditorConfig.class)
 public @interface EnableAuditor {}
