+++ auditor.outbound/src/main/java/com/github/tkaczenko/auditor/outbound/EnableOutboundAuditing.java
@@ -1,7 +1,7 @@
 package com.github.tkaczenko.auditor.outbound;
 
 import com.github.tkaczenko.auditor.core.EnableAuditor;
-import com.github.tkaczenko.auditor.outbound.config.AuditorConfiguration;
+import com.github.tkaczenko.auditor.outbound.config.AuditorConfig;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -16,7 +16,7 @@ import org.springframework.context.annotation.Import;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 @EnableAuditor
-@Import(AuditorConfiguration.class)
+@Import(AuditorConfig.class)
 public @interface EnableOutboundAuditing {
 
 }
