+++ auditor.starter/src/main/java/com/github/tkaczenko/auditor/starter/EnableHttpAuditing.java
@@ -1,6 +1,6 @@
 package com.github.tkaczenko.auditor.starter;
 
-import com.github.tkaczenko.auditor.starter.config.AuditorConfiguration;
+import com.github.tkaczenko.auditor.starter.config.AuditorConfig;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -14,5 +14,5 @@ import org.springframework.context.annotation.Import;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-@Import(AuditorConfiguration.class)
+@Import(AuditorConfig.class)
 public @interface EnableHttpAuditing {}
