+++ demo/src/main/java/com/github/tkaczenko/auditor/demo/config/FeignConfig.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class FeignConfiguration {
+public class FeignConfig {
 
   @Value("${feign.logging-level:BASIC}")
   private Logger.Level loggingLevel;
