+++ demo/src/main/java/com/github/tkaczenko/auditor/demo/config/JsonConfig.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class JsonConfiguration {
+public class JsonConfig {
 
   @Bean
   public ObjectMapper objectMapper() {
