+++ demo/src/main/java/com/github/tkaczenko/auditor/demo/config/PropertiesConfig.java
@@ -7,7 +7,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class PropertiesConfiguration {
+public class PropertiesConfig {
 
   @Data
   @Configuration
